Fixed bugs with db

+++ vagabond/db/migrate/20160506005215_remove_user_posts.rb
@@ -1,5 +1,7 @@
 class RemoveUserPosts < ActiveRecord::Migration
   def change
-    drop_table :user_posts
+    if ActiveRecord::Base.connection.tables.include?(:user_posts)
+      drop_table :user_posts
+    end
   end
 end
